Refactor selectedTake again and RecordScriptureFragment
